gaming: Explain what UMU_NO_RUNTIME does
Replace the ambigous note because the version written is obsolete and it can confuse users about which value
should be set. This also mentions that UMU_NO_RUNTIME is only meant for proton compiled against system
libraries. This is not limited to proton-cachyos, but also other proton versions that are compiled against
system libaries as opposed to the Steam Linux Runtime.

Signed-off-by: Eric Naim <dnaim@cachyos.org>